refactor: hhs_math
Changes made:

1. Renamed the function and variables to more descriptive names.
2. Added Doxygen style comments to the function.
3. Added comments to explain the logic of the code.

Please note that the original code seems to assume that the `curve` array is sorted in descending order of `lvl_mV`. If this is not the case, the function may not work as expected.

In this version, I've added a brief description at the beginning to explain what the function does. I've also added more detail to the parameter descriptions, and clarified the return value.